# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Patient service API we predicting patient data by name",
        "title": "Patient Service API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/v1",
    "paths": {
        "/patients": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Patients"
                ],
                "summary": "Get All Patients",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Name filter",
                        "name": "name",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Surname filter",
                        "name": "surname",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Patronymic filter",
                        "name": "patronymic",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response-controller_GetAllPatientsOutputResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Save Patiens and predict data by name",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Patients"
                ],
                "summary": "Create Patient",
                "parameters": [
                    {
                        "description": "Create Patient",
                        "name": "patient",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/converter.CreatePatientRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response-controller_CreateSinglePatientOutputResponse"
                        }
                    }
                }
            }
        },
        "/patients/{patientId}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Patients"
                ],
                "summary": "Get Patient",
                "parameters": [
                    {
                        "type": "string",
                        "description": "get patient by id",
                        "name": "patientId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response-controller_GetSinglePatientOutputResponse"
                        }
                    }
                }
            },
            "delete": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Patients"
                ],
                "summary": "Delete Patient",
                "parameters": [
                    {
                        "type": "string",
                        "description": "delete patient by id",
                        "name": "patientId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response-controller_DeleteSinglePatientOutputResponse"
                        }
                    }
                }
            },
            "patch": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Patients"
                ],
                "summary": "Update Patient",
                "parameters": [
                    {
                        "type": "string",
                        "description": "update patient by id",
                        "name": "patientId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Create Patient",
                        "name": "patient",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/converter.UpdatePatientRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response-controller_UpdateSinglePatientOutputResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "controller.CreateSinglePatientOutputResponse": {
            "type": "object",
            "properties": {
                "age": {
                    "type": "integer"
                },
                "country": {
                    "type": "string"
                },
                "gender": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "patronymic": {
                    "type": "string"
                },
                "surname": {
                    "type": "string"
                }
            }
        },
        "controller.DeleteSinglePatientOutputResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "patronymic": {
                    "type": "string"
                },
                "surname": {
                    "type": "string"
                }
            }
        },
        "controller.GetAllPatientsOutputResponse": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/response.PatientOutputResponse"
                    }
                },
                "pagination": {
                    "$ref": "#/definitions/response.Pagination"
                }
            }
        },
        "controller.GetSinglePatientOutputResponse": {
            "type": "object",
            "properties": {
                "age": {
                    "type": "integer"
                },
                "country": {
                    "type": "string"
                },
                "gender": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "patronymic": {
                    "type": "string"
                },
                "surname": {
                    "type": "string"
                }
            }
        },
        "controller.UpdateSinglePatientOutputResponse": {
            "type": "object",
            "properties": {
                "age": {
                    "type": "integer"
                },
                "country": {
                    "type": "string"
                },
                "gender": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "patronymic": {
                    "type": "string"
                },
                "surname": {
                    "type": "string"
                }
            }
        },
        "converter.CreatePatientRequest": {
            "type": "object",
            "required": [
                "name",
                "surname"
            ],
            "properties": {
                "name": {
                    "type": "string",
                    "maxLength": 15,
                    "minLength": 2
                },
                "patronymic": {
                    "type": "string",
                    "maxLength": 15,
                    "minLength": 2
                },
                "surname": {
                    "type": "string",
                    "maxLength": 15,
                    "minLength": 2
                }
            }
        },
        "converter.UpdatePatientRequest": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string",
                    "maxLength": 200,
                    "minLength": 2
                },
                "patronymic": {
                    "type": "string",
                    "maxLength": 15,
                    "minLength": 2
                },
                "surname": {
                    "type": "string",
                    "maxLength": 200,
                    "minLength": 2
                }
            }
        },
        "response.Pagination": {
            "type": "object",
            "properties": {
                "page": {
                    "type": "integer"
                },
                "perPage": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "response.PatientOutputResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "patronymic": {
                    "type": "string"
                },
                "surname": {
                    "type": "string"
                }
            }
        },
        "response.Response-controller_CreateSinglePatientOutputResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/controller.CreateSinglePatientOutputResponse"
                },
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
        "response.Response-controller_DeleteSinglePatientOutputResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/controller.DeleteSinglePatientOutputResponse"
                },
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
        "response.Response-controller_GetAllPatientsOutputResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/controller.GetAllPatientsOutputResponse"
                },
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
        "response.Response-controller_GetSinglePatientOutputResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/controller.GetSinglePatientOutputResponse"
                },
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
        "response.Response-controller_UpdateSinglePatientOutputResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/controller.UpdateSinglePatientOutputResponse"
                },
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                }
            }
        }
    }
}